feat: event bus cleanup
