+++ core/store/modules/order/mutations.ts
@@ -2,11 +2,8 @@ import Vue from 'vue'
 import { MutationTree } from 'vuex'
 import * as types from '../../mutation-types'
 import * as entities from '../../lib/entities'
-import EventBus from '../../lib/event-bus'
 import OrderState from './types/OrderState'
 import rootStore from '../../'
-
-declare var global: any
 
 const mutations: MutationTree<OrderState> = {
   /**
@@ -25,7 +22,7 @@ const mutations: MutationTree<OrderState> = {
       console.error(reason) // it doesn't work on SSR
     }, (err, resp) => {
       if (err) console.error(err)
-      EventBus.$emit('order/PROCESS_QUEUE', { config: rootStore.state.config }) // process checkout queue
+      Vue.prototype.$bus.$emit('order/PROCESS_QUEUE', { config: rootStore.state.config }) // process checkout queue
       console.info('Order placed, orderId = ' + orderId)
     }) // populate cache
   }
