+++ core/store/modules/product/mutations.ts
@@ -1,12 +1,12 @@
+import Vue from 'vue'
 import { MutationTree } from 'vuex'
 import * as types from '../../mutation-types'
-import EventBus from '../../lib/event-bus'
 import ProductState, { PagedProductList } from './types/ProductState'
 
 const mutations: MutationTree<ProductState> = {
   [types.CATALOG_UPD_RELATED] (state, { key, items }) {
     state.related[key] = items
-    EventBus.$emit('product-after-related', { key: key, items: items })
+    Vue.prototype.$bus.$emit('product-after-related', { key: key, items: items })
   },
   [types.CATALOG_ADD_CUSTOM_OPTION_VALIDATOR] (state, { validationRule, validatorFunction }) {
     state.custom_options_validators[validationRule] = validatorFunction
@@ -38,11 +38,11 @@ const mutations: MutationTree<ProductState> = {
   },
   [types.CATALOG_SET_PRODUCT_ORIGINAL] (state, product) {
     state.original = product
-    EventBus.$emit('product-after-original', { original: product })
+    Vue.prototype.$bus.$emit('product-after-original', { original: product })
   },
   [types.CATALOG_SET_PRODUCT_PARENT] (state, product) {
     state.parent = product
-    EventBus.$emit('product-after-parent', { parent: product })
+    Vue.prototype.$bus.$emit('product-after-parent', { parent: product })
   },
   [types.CATALOG_RESET_PRODUCT] (state, productOriginal) {
     state.current = productOriginal || {}
@@ -52,7 +52,7 @@ const mutations: MutationTree<ProductState> = {
     state.current_options = {color: [], size: []}
     state.current_bundle_options = {}
     state.current_custom_options = {}
-    EventBus.$emit('product-after-reset', { })
+    Vue.prototype.$bus.$emit('product-after-reset', { })
   },
   [types.CATALOG_UPD_GALLERY] (state, productGallery) {
     state.productGallery = productGallery
